Rename file to match rule
